Support custom module makefiles
Some special modules have post-install hooks (symlinks for example)
that must be run when the module is installed--these can't be installed
with a simple PRODUCT_COPY_FILES target.

This allows custom module makefiles in the device directory that are
named <module>.mk. The source file listed under LOCAL_SRC_FILES will be
extracted as usual, and the exact makefile will be appended to
Android.mk.

Some limitations right now (could be fixed when needed):

- custom modules are assumed to be normal files, not optimized bytecode
- only one file can be included under LOCAL_SRC_FILES

Signed-off-by: Preetam D'Souza <preetamjdsouza@gmail.com>